fix(team): split team workbench route chunk
