+++ web/src/pages/team/TeamThreadContainer.tsx
@@ -6,7 +6,7 @@ import {
   resolveChatMessageText,
   resolveThreadRootMessageIdFromPayload,
 } from "./page_helpers";
-import { buildTeamWorkspacePath } from "../team_page";
+import { buildTeamWorkspacePath } from "../../app_route_selection";
 import {
   createDisplayNameLookup,
   isHumanMailboxActor,
