+++ web/src/pages/team/use_team_member_agent_backfill_effect.ts
@@ -1,1 +1,1 @@
-export { useTeamMemberBackfillEffect as useTeamMemberAgentBackfillEffect } from "./use_team_member_backfill_effect";
+export { useTeamMemberBackfillEffect as useTeamMemberAgentBackfillEffect } from "../../team_member_backfill_effect";
